Refactor domain package structure to use `domain.model`
- Relocate `Movie`, `MovieDetail`, `MovieListType`, and `VideoStream` from `com.gabrielbmoro.moviedb.domain.entities` to `com.gabrielbmoro.moviedb.domain.model`.
- Move `HttpException` from the `domain` root package to the `domain.model` package and rename the file to `HttpException.kt`.
- Update all class imports across the `data`, `domain`, and feature modules to reflect the new package structure.
- Update package declarations and directory locations for corresponding tests in the domain module.
