+++ src/domain/src/commonMain/kotlin/com/gabrielbmoro/moviedb/domain/model/HttpException.kt
@@ -1,8 +1,8 @@
-package com.gabrielbmoro.moviedb.domain
+package com.gabrielbmoro.moviedb.domain.model
 
 data class HttpException(
     val statusCode: Int,
     val statusDescription: String,
     val requestUrl: String,
-    val requestMethod: String
+    val requestMethod: String,
 ) : IllegalStateException()
