+++ src/domain/src/commonMain/kotlin/com/gabrielbmoro/moviedb/domain/usecases/FavoriteMovieUseCase.kt
@@ -1,7 +1,7 @@
 package com.gabrielbmoro.moviedb.domain.usecases
 
 import com.gabrielbmoro.moviedb.domain.MoviesRepository
-import com.gabrielbmoro.moviedb.domain.entities.Movie
+import com.gabrielbmoro.moviedb.domain.model.Movie
 import org.koin.core.annotation.Factory
 import org.koin.core.annotation.Provided
