+++ src/domain/src/commonTest/kotlin/com/gabrielbmoro/moviedb/domain/model/MoviesListTypeTest.kt
@@ -1,4 +1,4 @@
-package com.gabrielbmoro.moviedb.domain.entities
+package com.gabrielbmoro.moviedb.domain.model
 
 import kotlin.test.Test
 import kotlin.test.assertEquals
