+++ src/feature/feature-movies/src/commonTest/kotlin/com/gabrielbmoro/moviedb/movies/components/MoviesHandlerTest.kt
@@ -1,7 +1,7 @@
 package com.gabrielbmoro.moviedb.movies.components
 
 import MoviesHandler
-import com.gabrielbmoro.moviedb.domain.entities.Movie
+import com.gabrielbmoro.moviedb.domain.model.Movie
 import com.gabrielbmoro.moviedb.movies.fakes.FakeRepository
 import com.gabrielbmoro.moviedb.movies.ui.screens.movies.FilterType
 import kotlinx.coroutines.test.runTest
