+++ src/feature/feature-movies/src/commonTest/kotlin/com/gabrielbmoro/moviedb/movies/fakes/FakeRepository.kt
@@ -1,9 +1,9 @@
 package com.gabrielbmoro.moviedb.movies.fakes
 
 import com.gabrielbmoro.moviedb.domain.MoviesRepository
-import com.gabrielbmoro.moviedb.domain.entities.Movie
-import com.gabrielbmoro.moviedb.domain.entities.MovieDetail
-import com.gabrielbmoro.moviedb.domain.entities.VideoStream
+import com.gabrielbmoro.moviedb.domain.model.Movie
+import com.gabrielbmoro.moviedb.domain.model.MovieDetail
+import com.gabrielbmoro.moviedb.domain.model.VideoStream
 
 class FakeRepository(
     private val getMoviesFromFilterError: Throwable? = null,
