+++ src/feature/feature-search/src/commonTest/kotlin/com/gabrielbmoro/moviedb/search/ui/screens/search/FakeRepository.kt
@@ -1,9 +1,9 @@
 package com.gabrielbmoro.moviedb.search.ui.screens.search
 
 import com.gabrielbmoro.moviedb.domain.MoviesRepository
-import com.gabrielbmoro.moviedb.domain.entities.Movie
-import com.gabrielbmoro.moviedb.domain.entities.MovieDetail
-import com.gabrielbmoro.moviedb.domain.entities.VideoStream
+import com.gabrielbmoro.moviedb.domain.model.Movie
+import com.gabrielbmoro.moviedb.domain.model.MovieDetail
+import com.gabrielbmoro.moviedb.domain.model.VideoStream
 
 class FakeRepository : MoviesRepository {
     lateinit var searchResult: List<Movie>
