+++ src/feature/feature-wishlist/src/commonMain/kotlin/com/gabrielbmoro/moviedb/wishlist/ui/screens/wishlist/WishlistViewModel.kt
@@ -1,7 +1,7 @@
 package com.gabrielbmoro.moviedb.wishlist.ui.screens.wishlist
 
 import com.gabrielbmoro.moviedb.domain.MoviesRepository
-import com.gabrielbmoro.moviedb.domain.entities.Movie
+import com.gabrielbmoro.moviedb.domain.model.Movie
 import com.gabrielbmoro.moviedb.domain.usecases.FavoriteMovieUseCase
 import com.gabrielbmoro.moviedb.platform.viewmodel.BaseViewModel
 import com.gabrielbmoro.moviedb.wishlist.ui.widgets.MovieCardInfo
